+++ src/settings.ts
@@ -1,28 +1,23 @@
 'use strict';
 
-import vscode = require("vscode");
+import vscode = require('vscode');
 
-export enum PuppetInstallType{
-  PDK    = "pdk",
-  PUPPET = "agent",
-  AUTO   = "auto",
+export enum PuppetInstallType {
+  PDK = 'pdk',
+  PUPPET = 'agent',
+  AUTO = 'auto'
 }
 
 export enum ProtocolType {
   UNKNOWN = '<unknown>',
   STDIO = 'stdio',
-  TCP = 'tcp',
-  DOCKER = "docker"
+  TCP = 'tcp'
 }
 
 export enum ConnectionType {
   Unknown,
   Local,
   Remote
-}
-
-export interface IEditorServiceDockerSettings {
-  imageName?: string;
 }
 
 export interface IEditorServiceTCPSettings {
@@ -40,7 +35,6 @@ export interface IEditorServicePuppetSettings {
 
 export interface IEditorServiceSettings {
   debugFilePath?: string;
-  docker?: IEditorServiceDockerSettings;
   enable?: boolean;
   featureFlags?: string[];
   loglevel?: string;
@@ -83,7 +77,7 @@ export interface ISettings {
   pdk?: IPDKSettings;
 }
 
-const workspaceSectionName = "puppet";
+const workspaceSectionName = 'puppet';
 
 /**
  * Safely query the workspace configuration for a nested setting option.  If it, or any part of the setting
@@ -91,18 +85,22 @@ const workspaceSectionName = "puppet";
  * @param workspaceConfig The VScode workspace configuration to query
  * @param indexes         An array of strings defining the setting path, e.g. The setting 'a.b.c' would pass indexes of ['a','b','c']
  */
-function getSafeWorkspaceConfig(workspaceConfig: vscode.WorkspaceConfiguration, indexes:string[] ): any {
-  if (indexes.length <= 0) { return undefined; }
+function getSafeWorkspaceConfig(workspaceConfig: vscode.WorkspaceConfiguration, indexes: string[]): any {
+  if (indexes.length <= 0) {
+    return undefined;
+  }
 
   let index: string = indexes.shift();
   let result: Object = workspaceConfig[index];
-  while ((indexes.length > 0) && (result !== undefined) ) {
+  while (indexes.length > 0 && result !== undefined) {
     index = indexes.shift();
     result = result[index];
   }
 
   // A null settings is really undefined.
-  if (result === null) { return undefined; }
+  if (result === null) {
+    return undefined;
+  }
 
   return result;
 }
@@ -117,40 +115,58 @@ export function legacySettings(): Map<string, Object> {
   let value: Object = undefined;
 
   // puppet.editorService.modulePath
-  value = getSafeWorkspaceConfig(workspaceConfig, ['editorService','modulePath']);
-  if (value !== undefined) { settings.set("puppet.editorService.modulePath", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['editorService', 'modulePath']);
+  if (value !== undefined) {
+    settings.set('puppet.editorService.modulePath', value);
+  }
 
   // puppet.languageclient.minimumUserLogLevel
-  value = getSafeWorkspaceConfig(workspaceConfig, ['languageclient','minimumUserLogLevel']);
-  if (value !== undefined) { settings.set("puppet.languageclient.minimumUserLogLevel", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['languageclient', 'minimumUserLogLevel']);
+  if (value !== undefined) {
+    settings.set('puppet.languageclient.minimumUserLogLevel', value);
+  }
 
   // puppet.languageclient.protocol
-  value = getSafeWorkspaceConfig(workspaceConfig, ['languageclient','protocol']);
-  if (value !== undefined) { settings.set("puppet.languageclient.protocol", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['languageclient', 'protocol']);
+  if (value !== undefined) {
+    settings.set('puppet.languageclient.protocol', value);
+  }
 
   // puppet.languageserver.address
-  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver','address']);
-  if (value !== undefined) { settings.set("puppet.languageserver.address", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver', 'address']);
+  if (value !== undefined) {
+    settings.set('puppet.languageserver.address', value);
+  }
 
   // puppet.languageserver.debugFilePath
-  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver','debugFilePath']);
-  if (value !== undefined) { settings.set("puppet.languageserver.debugFilePath", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver', 'debugFilePath']);
+  if (value !== undefined) {
+    settings.set('puppet.languageserver.debugFilePath', value);
+  }
 
   // puppet.languageserver.filecache.enable
-  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver','filecache','enable']);
-  if (value !== undefined) { settings.set("puppet.languageserver.filecache.enable", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver', 'filecache', 'enable']);
+  if (value !== undefined) {
+    settings.set('puppet.languageserver.filecache.enable', value);
+  }
 
   // puppet.languageserver.port
-  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver','port']);
-  if (value !== undefined) { settings.set("puppet.languageserver.port", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver', 'port']);
+  if (value !== undefined) {
+    settings.set('puppet.languageserver.port', value);
+  }
 
   // puppet.languageserver.timeout
-  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver','timeout']);
-  if (value !== undefined) { settings.set("puppet.languageserver.timeout", value); }
+  value = getSafeWorkspaceConfig(workspaceConfig, ['languageserver', 'timeout']);
+  if (value !== undefined) {
+    settings.set('puppet.languageserver.timeout', value);
+  }
 
   // puppet.puppetAgentDir
   value = getSafeWorkspaceConfig(workspaceConfig, ['puppetAgentDir']);
-  if (value !== undefined) { settings.set("puppet.puppetAgentDir", value); }
+  if (value !== undefined) {
+    settings.set('puppet.puppetAgentDir', value);
+  }
 
   return settings;
 }
@@ -161,7 +177,7 @@ export function DefaultWorkspaceSettings(): ISettings {
     editorService: {
       enable: true,
       featureFlags: [],
-      loglevel: "normal",
+      loglevel: 'normal',
       protocol: ProtocolType.STDIO,
       timeout: 10
     },
@@ -169,19 +185,19 @@ export function DefaultWorkspaceSettings(): ISettings {
       enable: true
     },
     hover: {
-      showMetadataInfo: true,
+      showMetadataInfo: true
     },
     installDirectory: undefined,
     installType: PuppetInstallType.AUTO,
     lint: {
-      enable: true,
+      enable: true
     },
     notification: {
-      nodeGraph: "messagebox",
-      puppetResource: "messagebox"
+      nodeGraph: 'messagebox',
+      puppetResource: 'messagebox'
     },
     pdk: {
-      checkVersion: true,
+      checkVersion: true
     }
   };
 }
@@ -192,28 +208,36 @@ export function SettingsFromWorkspace(): ISettings {
 
   // TODO: What if the wrong type is passed through? will it blow up?
   let settings = {
-    editorService: workspaceConfig.get<IEditorServiceSettings>("editorService", defaults.editorService),
-    format: workspaceConfig.get<IFormatSettings>("format",  defaults.format),
-    hover: workspaceConfig.get<IHoverSettings>("hover",  defaults.hover),
-    installDirectory: workspaceConfig.get<string>("installDirectory",  defaults.installDirectory),
-    installType: workspaceConfig.get<PuppetInstallType>("installType",  defaults.installType),
-    lint: workspaceConfig.get<ILintSettings>("lint",  defaults.lint),
-    notification: workspaceConfig.get<INotificationSettings>("notification", defaults.notification),
-    pdk: workspaceConfig.get<IPDKSettings>("pdk", defaults.pdk)
+    editorService: workspaceConfig.get<IEditorServiceSettings>('editorService', defaults.editorService),
+    format: workspaceConfig.get<IFormatSettings>('format', defaults.format),
+    hover: workspaceConfig.get<IHoverSettings>('hover', defaults.hover),
+    installDirectory: workspaceConfig.get<string>('installDirectory', defaults.installDirectory),
+    installType: workspaceConfig.get<PuppetInstallType>('installType', defaults.installType),
+    lint: workspaceConfig.get<ILintSettings>('lint', defaults.lint),
+    notification: workspaceConfig.get<INotificationSettings>('notification', defaults.notification),
+    pdk: workspaceConfig.get<IPDKSettings>('pdk', defaults.pdk)
   };
 
   /**
    * Legacy Workspace Settings
-   * 
-   * Retrieve deprecated settings and apply them to the settings.  This is only needed as a helper and should be 
+   *
+   * Retrieve deprecated settings and apply them to the settings.  This is only needed as a helper and should be
    * removed a version or two later, after the setting is deprecated.
    */
 
-   // Ensure that object types needed for legacy settings exists
-  if (settings.editorService === undefined) { settings.editorService = {}; }
-  if (settings.editorService.featureFlags === undefined) { settings.editorService.featureFlags = []; }
-  if (settings.editorService.puppet === undefined) { settings.editorService.puppet = {}; }
-  if (settings.editorService.tcp === undefined) { settings.editorService.tcp = {}; }
+  // Ensure that object types needed for legacy settings exists
+  if (settings.editorService === undefined) {
+    settings.editorService = {};
+  }
+  if (settings.editorService.featureFlags === undefined) {
+    settings.editorService.featureFlags = [];
+  }
+  if (settings.editorService.puppet === undefined) {
+    settings.editorService.puppet = {};
+  }
+  if (settings.editorService.tcp === undefined) {
+    settings.editorService.tcp = {};
+  }
 
   // Retrieve the legacy settings
   const oldSettings: Map<string, Object> = legacySettings();
@@ -221,43 +245,44 @@ export function SettingsFromWorkspace(): ISettings {
   // Translate the legacy settings into the new setting names
   for (const [settingName, value] of oldSettings) {
     switch (settingName) {
-
-      case "puppet.editorService.modulePath": // --> puppet.editorService.puppet.modulePath
+      case 'puppet.editorService.modulePath': // --> puppet.editorService.puppet.modulePath
         settings.editorService.puppet.modulePath = <string>value;
         break;
 
-      case "puppet.languageclient.minimumUserLogLevel": // --> puppet.editorService.loglevel
+      case 'puppet.languageclient.minimumUserLogLevel': // --> puppet.editorService.loglevel
         settings.editorService.loglevel = <string>value;
         break;
 
-      case "puppet.languageclient.protocol": // --> puppet.editorService.protocol
+      case 'puppet.languageclient.protocol': // --> puppet.editorService.protocol
         settings.editorService.protocol = <ProtocolType>value;
         break;
 
-      case "puppet.languageserver.address": // --> puppet.editorService.tcp.address
+      case 'puppet.languageserver.address': // --> puppet.editorService.tcp.address
         settings.editorService.tcp.address = <string>value;
         break;
 
-      case "puppet.languageserver.debugFilePath": // --> puppet.editorService.debugFilePath
+      case 'puppet.languageserver.debugFilePath': // --> puppet.editorService.debugFilePath
         settings.editorService.debugFilePath = <string>value;
         break;
 
-      case "puppet.languageserver.filecache.enable": // --> puppet.editorService.featureflags['filecache']
-        if (value === true) { settings.editorService.featureFlags.push("filecache"); }
+      case 'puppet.languageserver.filecache.enable': // --> puppet.editorService.featureflags['filecache']
+        if (value === true) {
+          settings.editorService.featureFlags.push('filecache');
+        }
         break;
 
-      case "puppet.languageserver.port": // --> puppet.editorService.tcp.port
+      case 'puppet.languageserver.port': // --> puppet.editorService.tcp.port
         settings.editorService.tcp.port = <number>value;
         break;
 
-      case "puppet.languageserver.timeout": // --> puppet.editorService.timeout
+      case 'puppet.languageserver.timeout': // --> puppet.editorService.timeout
         settings.editorService.timeout = <number>value;
         break;
 
-      case "puppet.puppetAgentDir": // --> puppet.installDirectory
+      case 'puppet.puppetAgentDir': // --> puppet.installDirectory
         settings.installDirectory = <string>value;
         break;
-      }
+    }
   }
 
   return settings;